+++ users/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v2.0.0] users
 ### Releases
 - [NPM v2.0.0] https://www.npmjs.com/package/@wingspan/users/v/2.0.0 - users
+
+## 2023-10-20 20:44:55
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.103.0 (2.168.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] users
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@wingspan/users/v/2.1.0 - users